add to paarser get_resource_url_name
