+++ page_loader/parser.py
@@ -1,6 +1,7 @@
 import re
 from urllib.parse import urlparse
 import os
+from typing import Tuple
 
 RE_URL = (
     r'^(https?:\/\/)?([\da-z\.-]+)\.([a-z\.]+)([\/\w \.-]*)*\/?(\??(.?)*)$'
@@ -43,6 +44,29 @@ def get_url_name(url: dict) -> str:
     parsed_path, _ = os.path.splitext(url['path'])
     without_scheme_url = url['netloc'] + parsed_path
     return normalize_name(without_scheme_url)
+
+
+def get_resource_url_name(
+    value: str,
+    url: dict
+) -> Tuple[str, str]:
+    """Generate the file name by url."""
+    parsed_value_url = urlparse(value)
+
+    parsed_path, extention = os.path.splitext(parsed_value_url.path.strip('/'))
+    parsed_value_path = normalize_name(parsed_path) + extention
+
+    if not parsed_value_url.scheme:
+        target_address = '{scheme}{netloc}{path}?{query}'.format(
+            scheme=url['scheme'],
+            netloc=url['netloc'],
+            path=parsed_value_url.path,
+            query=parsed_value_url.query,
+        )
+    else:
+        target_address = value
+
+    return (target_address, parsed_value_path)
 
 
 def normalize_name(name: str) -> str:
